Refactor, split in files
